cleanup: txtp
